amount float

+++ backend/app/modules/finance/models.py
@@ -1,5 +1,5 @@
 from app.core.database import Base
-from sqlalchemy import Boolean, Column, ForeignKey, Integer, String
+from sqlalchemy import Boolean, Column, ForeignKey, Integer, String, Float
 from sqlalchemy.orm import relationship
 
 
@@ -9,7 +9,7 @@ class Account(Base):
     id = Column(Integer, primary_key=True)
     userId = Column(Integer, ForeignKey("users.id"))
     name = Column(String)
-    amount = Column(Integer)
+    amount = Column(Float)
     pinCode = Column(String)
     is_active = Column(Boolean, default=True, nullable=False)
     user = relationship("User")
